requirements.txt
flask
flask-restful
flask-sqlalchemy
python-dotenv
pyjwt
requests
waitress